+++ services/src/main/java/ar/edu/itba/paw/services/ImageServiceImpl.java
@@ -19,7 +19,7 @@ public class ImageServiceImpl implements ImageService {
     }
 
     @Override
-    public Optional<Image> uploadImage(byte[] photoBlob, Integer imageContentLength, String imageContentType) {
+    public Optional<Image> uploadImage(byte[] photoBlob, long imageContentLength, String imageContentType) {
         return imageDao.uploadImage(photoBlob, imageContentLength, imageContentType);
     }
 }
